api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
redis==5.0.3
kafka-python==2.0.2